+++ jest.config.js
@@ -22,6 +22,8 @@ module.exports = {
     '\\.(css|scss)$': '<rootDir>/script/jest/styleMock',
     '^~styleguide/blocks(.*)$':
       '<rootDir>/packages/styleguide/.storybook/components$1',
+    '^@codecademy\\/storybook-addon-variance(.*)$':
+      '<rootDir>/packages/styleguide/.storybook/addons/system/components$1',
     '^@codecademy\\/gamut$': '<rootDir>/packages/gamut/src',
     '^@codecademy\\/gamut-labs$': '<rootDir>/packages/gamut-labs/src',
     '^@codecademy\\/gamut-styles$': '<rootDir>/packages/gamut-styles/src',
@@ -31,6 +33,7 @@ module.exports = {
     '^@codecademy\\/gamut-illustrations$':
       '<rootDir>/packages/gamut-illustrations/src',
     '^@codecademy\\/macros$': '<rootDir>/packages/macros',
+    '^@codecademy\\/gamut-patterns$': '<rootDir>/packages/gamut-patterns/src',
   },
   testPathIgnorePatterns: ['/node_modules/', '/dist/'],
   transform: {
@@ -53,6 +56,7 @@ module.exports = {
     '/tmp/',
     '/example/',
     '/typings/',
+    '/.storybook/',
   ],
   reporters: process.env.CI ? ['default', 'jest-junit'] : ['default'],
   coverageReporters: ['json', 'text', 'clover'],