+++ packages/gamut-styles/src/ColorMode.tsx
@@ -1,9 +1,8 @@
-import { serializeTokens, variance } from '@codecademy/variance';
+import { serializeTokens, StyleProps, variance } from '@codecademy/variance';
 import { CSSObject, Theme, ThemeProvider, useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
-import { HTMLAttributes } from 'enzyme';
 import { mapValues } from 'lodash';
-import React, { useMemo } from 'react';
+import React, { ComponentProps, forwardRef, useMemo } from 'react';
 
 import {
   color,
@@ -15,8 +14,14 @@ import {
 } from './variance/props';
 import { styledConfig } from './variance/utils';
 
+export type Colors = keyof Theme['colors'];
+export type ColorModeConfig = Theme['modes'];
+export type ColorModes = keyof ColorModeConfig;
+export type ColorModeShape = ColorModeConfig[ColorModes];
+export type ColorAlias = keyof ColorModeShape;
+
 export type ColorModeProps = {
-  mode: keyof Theme['colorModes']['modes'];
+  mode: ColorModes;
   bg?: keyof Theme['colors'];
   className?: string;
 };
@@ -30,28 +35,34 @@ export const providerProps = variance.compose(
   space
 );
 
-export type ProviderProps = Parameters<typeof providerProps>[0];
-
-export interface VariableProviderProps
-  extends Omit<HTMLAttributes, keyof ProviderProps>,
-    ProviderProps {
-  variables?: CSSObject;
+export function useColorModes(): [
+  ColorModes,
+  ColorModeShape,
+  ColorModeConfig,
+  (color: Colors) => string
+] {
+  const { mode, modes, _getColorValue: getColorValue } = useTheme() || {};
+  return [mode, modes?.[mode], modes, getColorValue];
 }
 
-export const VariableProvider = styled(
-  'div',
-  styledConfig
-)<VariableProviderProps>(({ variables }) => variables, providerProps);
+export function useCurrentMode(mode?: ColorModes) {
+  const [activeMode] = useColorModes();
+  return mode ?? activeMode;
+}
 
-export const ColorMode: React.FC<ColorModeProps & ProviderProps> = ({
-  mode,
-  ...rest
-}) => {
+export const VariableProvider = styled('div', styledConfig)<
+  StyleProps<typeof providerProps> & {
+    variables?: CSSObject;
+    alwaysSetVariables?: boolean;
+  }
+>(({ variables }) => variables, providerProps);
+
+export const ColorMode = forwardRef<
+  HTMLDivElement,
+  ColorModeProps & ComponentProps<typeof VariableProvider>
+>(({ mode, alwaysSetVariables, ...rest }, ref) => {
   const theme = useTheme();
-  const {
-    colorModes: { modes, active },
-    colors,
-  } = theme;
+  const { modes, mode: active, colors } = theme;
   const { variables } = useMemo(
     () =>
       serializeTokens(
@@ -62,12 +73,23 @@ export const ColorMode: React.FC<ColorModeProps & ProviderProps> = ({
     [colors, mode, modes, theme]
   );
   if (active === mode) {
-    return <VariableProvider {...rest} />;
+    return (
+      <VariableProvider
+        {...rest}
+        ref={ref}
+        variables={alwaysSetVariables ? variables : undefined}
+      />
+    );
   }
 
   return (
-    <ThemeProvider theme={{ colorModes: { modes, active: mode } }}>
-      <VariableProvider variables={variables} textColor="text" {...rest} />
+    <ThemeProvider theme={{ mode }}>
+      <VariableProvider
+        variables={variables}
+        textColor="text"
+        {...rest}
+        ref={ref}
+      />
     </ThemeProvider>
   );
-};
+});