+++ packages/gamut/src/Form/FormGroupLabel.tsx
@@ -51,7 +51,7 @@ const labelSizeVariants = variant({
 const disabledLabelStyle = ({ disabled }: disabledLabelStyleProps) => {
   if (disabled) {
     return css`
-      color: ${theme.colors[`gray-400`]};
+      color: ${theme.colors['gray-600']};
     `;
   }
 };
@@ -97,7 +97,7 @@ export const FormGroupLabel: React.FC<FormGroupLabelProps> = ({
           <StyledToolTip
             {...tooltip}
             alignment="bottom-right"
-            target={<MiniInfoOutlineIcon height="0.8rem" width="0.8rem" />}
+            target={<MiniInfoOutlineIcon size="0.8rem" />}
           />
         </StyledToolTipContainer>
       )}