+++ packages/gamut/src/NotificationList/typings.ts
@@ -1,5 +1,3 @@
-import { iconMap } from '../deprecated/Icon/iconMap';
-
 export type NotificationIconSettings = {
   backgroundColor?: string;
   color?: string;
@@ -9,7 +7,7 @@ export type Notification = {
   date: string;
   id: string;
   iconSettings?: NotificationIconSettings;
-  iconSlug?: keyof typeof iconMap;
+  iconSlug?: string;
   imageUrl?: string;
   link?: string;
   text: string;